working script, WIP README

+++ upgrade/v2.5/script/upgrade.sh
@@ -14,9 +14,15 @@ echo "Migrating IPPools..."
 retval=$?
 if [ $retval == 0 ];
 then
-    echo "Successfully got the Global BGP Peers:"
-    cat ippool.yaml
-    /sbin/calicoctl-v1.5 apply -f ippool.yaml
+    if [ -f ippool.yaml ]; then
+        echo "Successfully got the IPPools:"
+        cat ippool.yaml
+        /sbin/calicoctl-v1.5 apply -f ippool.yaml
+    else
+       echo "No IPPools found to migrate."
+    fi
+else
+    echo "Failed to get IPPools"
 fi
 
 # Get BGPPeers list with old calicoctl (from TPR) and save it as a yaml file.
@@ -26,9 +32,15 @@ echo "Migrating BGPPeers..."
 retval=$?
 if [ $retval == 0 ];
 then
-    echo "Successfully got the Global BGP Peers:"
-    cat bgppeer.yaml
-    /sbin/calicoctl-v1.5 apply -f bgppeer.yaml
+    if [ -f bgppeer.yaml ]; then
+        echo "Successfully got the Global BGP Peers:"
+        cat bgppeer.yaml
+        /sbin/calicoctl-v1.5 apply -f bgppeer.yaml
+    else
+       echo "No BGPPeers found to migrate."
+    fi
+else
+    echo "Failed to get Global BGP Peers"
 fi
 
 
@@ -37,16 +49,22 @@ fi
 # rename resource kind from 'GlobalConfig' to 'GlobalFelixConfig' and save it in a new yaml file.
 # Apply the modified yaml file.
 echo "Migrating GlobalFelixConfig..."
-/sbin/kubectl get globalconfig --all-namespaces -o yaml > cat tpr-felixconfig.yaml
+/sbin/kubectl get globalconfig --all-namespaces -o yaml > tpr-felixconfig.yaml
 retval=$?
 if [ $retval == 0 ];
 then
-    echo "Successfully got the Felix Config:"
-    cat tpr-felixconfig.yaml
-    cat tpr-felixconfig.yaml | sed '/apiVersion/s/projectcalico\.org\/v1/crd\.projectcalico\.org\/v1/g' | sed '/kind/s/GlobalConfig/GlobalFelixConfig/g' > crd-felixconfig.yaml
-    echo "Successfully renamed the resource kind"
-    cat crd-felixconfig.yaml
-    -/sbin/kubectl apply -f crd-felixconfig.yaml
+    if [ -f tpr-felixconfig.yaml ]; then
+        echo "Successfully got the Felix Config:"
+        cat tpr-felixconfig.yaml
+        cat tpr-felixconfig.yaml | sed '/apiVersion/s/projectcalico\.org\/v1/crd\.projectcalico\.org\/v1/g' | sed '/kind/s/GlobalConfig/GlobalFelixConfig/g' > crd-felixconfig.yaml
+        echo "Successfully renamed the resource kind"
+        cat crd-felixconfig.yaml
+        /sbin/kubectl apply -f crd-felixconfig.yaml
+    else
+       echo "No GlobalFelixConfig found to migrate."
+    fi
+else
+    echo "Failed to get Global Felix config"
 fi
 
 
@@ -58,12 +76,18 @@ echo "Migrating GlobalBGPConfig..."
 retval=$?
 if [ $retval == 0 ];
 then
-    echo "Successfully got the BGP Config:"
-    cat tpr-bgpconfig.yaml
-    cat tpr-bgpconfig.yaml | sed '/apiVersion/s/projectcalico\.org\/v1/crd\.projectcalico\.org\/v1/g' > crd-bgpconfig.yaml
-    echo "Successfully renames the resource kind"
-    cat crd-bgpconfig.yaml
-    -/sbin/kubectl apply -f crd-bgpconfig.yaml
+    if [ -f tpr-bgpconfig.yaml ]; then
+        echo "Successfully got the BGP Config:"
+        cat tpr-bgpconfig.yaml
+        cat tpr-bgpconfig.yaml | sed '/apiVersion/s/projectcalico\.org\/v1/crd\.projectcalico\.org\/v1/g' > crd-bgpconfig.yaml
+        echo "Successfully renamed the resource kind"
+        cat crd-bgpconfig.yaml
+        /sbin/kubectl apply -f crd-bgpconfig.yaml
+    else
+       echo "No GlobalBGPConfig found to migrate."
+    fi
+else
+    echo "Failed to get BGP config"
 fi
